initial Package Commit
Added packages that will be used for production.

The Oculus integration package will likely only be used to get properly animated hands if I can even figure those out. Aside from that, it will likely not be used further

Photon is used for multiplayer

XR interaction toolkit will be used to create the controls for the games