Local Folder Restructure
I was getting weird errors and one of the messages that I read on them essentially said that my files were bvuried too deep in my directory. Removing project folder from git repo to make a new one that is not buried as deep.

I will be using a different render pipeline in unity and want to avoid any struggle that may come in changing pipelines so a new project is being made once again