VR Setup
Updated Unity Settings so that builds can be sent to Oculus.

An error would appear when trying to build the game when Project Settings > XR Plug-in Management > Plug-in Providers was set to OpenXR. Somthing about a depracated option in the unity options(Android.enableR8). No fixes online were found that kept the gradle version of android at the current version.