Wall Timers
Added Sliders that function as timers to the walls of the room. One side has the time left along with a slider while the others only have the timer. Also made it  so sliders and text turn red in the last 10 seconds

changed Start() function in Network manager to call the ConnectToServer() function. The two did the same thing to begin with