Multiplayer Gun Tracking
Began updating scripts to track two players in a scene at once. Network Player now has a "reference" which is sent to the guns they are holding then to bullets fired.

Fixed a bug where the bullet was colliding with the gun that fired it and immediately deactivating
There's currently a bug where the player teleports more than once after enabling teleport. Either change teleport to permanently be on or fix bug