Further Updates to Target version 1
Target were hard to distinguish between. Targets are all supposed to have an index that decides what player can shoot them. Updated the hit target script so that the target color scheme is changed depending on which index it is set on. 0(both players) is purple, 1(player 1) is blue, and 2(player 2) is red.

Lowered intensity of post processing to make lights a little more bearable.

Started cleaning and polishing code.