Updated hit targets
updated hit targets to change target types spawned depending on which scene it's on. Will be changed soon to use a different  hit target script

Updated target controller so that spawning interval can be changed from editor

Game manager immediately starts spawning targets in main menu scene

All targets are removed when timer is completed(This crashed unity last time it was tested. Will prolly remove it if it continues to crash the game)